new exercise - this one includes an image, and uses the "inexact" solution type
This question had been entered in coursera as needing a numeric answer
in the range of 1.20 to 1.21.  I initially set this up using the
inexact data type with 1.205 set as the answer, and a max error of
.005.  This didn't quite work, since it marked 1.20 as wrong (not sure
why).  So I used a max error of .0051 to get around the problem instead
of trying to understand what went wrong.