fixed show-definitions (thanks Jason!)
The examples I found of using show-definitions in the Khan files all
included the <div class="definition"> just after the <p> containing the
item to be defined, but I couldn't get it to work that way for me.  To
get it to work, I needed to include it in the same <p> or <div> as the
item, above it.  Note that though it doesn't need to be in a hint (it
could e.g. be in a question), you need to include the module sub-hints
to get it to work.  This file includes 2 examples of definitions, one
in a hint, and one in the question text.