Add redirect to /nlp/Fall2012/.
Now if you go to / or /nlp you get redirected to the one course we're
prototyping up now, /nlp/Fall2012.  Much nicer for the overall flow,
especially login.

To do this I had to park the c2g.home somewhere, so I created a bogus
courses URL that we can build out later (or ignore).