Fix problem set link on home page.
This fixes the links under "Problem Sets" on the left pane of the
homepage.  The link was using to the pset.id, meaning that we got links
like this that went nowhere:

http://localhost:8000/nlp/Fall2012/problemsets/1

Now using the title the link is this, which works with the "problemsets"
area now.  Maybe not the right thing long term, but in that case, we can
fix it later.

http://localhost:8000/nlp/Fall2012/problemsets/P1

Needs a code review from Kelvin.