Piazza integration: use template for outer frame
The main view function now calls a template that renders the navigation
elements (mostly coming from course_base.html) but then has an iframe
that references the second URL, to be rendered by piazza_iframe().  This
is now the function that makes the LTI call and writes the result back
to the iframe.

Still not working yet, probably due to relative paths on the Piazza
side...