Remove placeholder course views.
Rename course.views.view to course.views.main.  A view named "view" is
confusing, especially when it's not the only view.

Removed: prereqs, staff, mine, current, staff.  Those can be re-added
if/when make sense.

Chanve to course_base needed so that the global nav bar points to the
new names.