Initial checkin of Class2Go cookbooks
Initial cookbooks for managing class2go production and staging.  The
ones that are specific to managing the class2go setup are prefixed with
"class2go", the others are general.  I wrote the "gdata" one, which is
just packaging up the google data distribution, but since it's not
specific to our environment per se, it's got a non-c2g name.