add class2go-python recipe for installing dependency packages
This is where we will install packages that we depend on.  Currently
just django-storages and boto, but will be more later.  We've gone along
this long partially just because the base Bitnami image we're using
already picks up other things we depend on, like south.

Note that this didn't work with pip, caused all kinds of weird errors,
so fell back to easy_install.  Comment added.