Simpler image differencing to remove scipy/numpy dependency (#470)
The old way used some fancy math to compare images -- convert to
grayscale, normalize, etc.  To do that required the numpy/scipy
libraries that are a pain to get working on different machines (macs,
non-ubuntu AMI's).

This just does a simpler RMS calculation over the image.  Empirically
this seems to notice about the same number of differences, ie. cause the
same granularity of key frames to show up.

The thresholds will be different now.  15 was a good one before,
now 1000.  Again, determined empirically, and on just a couple of
videos (should look at more).  Force frame extraction to always
output 640x480 sized frames to make this as consistent as possible.

Also the logic for running locally was broken.  Fixed that too.