Move kelvinator from util proj back under main (#472)
Originally I thought that we would want the apps that would run on the
util machine to be in their own Django project. But moved into the
main Django project for two reasons:

1. Database - util apps will want to read/write from our shared database
   along the way, eg. to write back video extract manifests, or to read
   message bodies.  So they should share database settings and (more
   importantly) models.

2. Object dispatch - since we're using celery (and more specifically,
   django's bindings to it) we need to dispatch to @tasks, and not
   apparent how that would work across Django projects.

Putting in one project is fine.  We may just not want to fire up that
app on our appservers.