parameterized sophi-python; redo how patch is done (#519)
Previous way of doing patch was super fragile, with a hardcoded path
that was never going to stay stable.  Now I clevely :) figure out where
the storages module is installed, and then use that for the patch.

Requires running a script instead of just a command, so chose to do that
with a here-script.

Also using the easy_install_package chef resource instead of just
calling execute on the command.  More robust.