Parameterize celery settings: timeout, concurrency (#781)
We don't want to have to do code changes to change this stuff.
Parameterize and put defaults in environment file.  We could even do
per-machine overrides or such if we want to manage these settings more
carefully in the future.

Concurrency back to 8 in all the environment files, FWIW, so we can do
email blasts.  Will set back to 2 to for normal operations
(kelvination).

Tested on util1.dev, after running we got the right things in
/etc/defaults/celeryd:

> # Extra arguments to celeryd
> CELERYD_OPTS="--time-limit=7200 --concurrency=8"