Separate kelvinator and resize tasks onto "long" queue (#1135)
Use celery router to put long-running tasks on a separate queue.  So now
we will see "prod-celery" and "prod-long" queue's in SQS.

Configure workers so instead of one worker named just "w1" now we
will have "default" and "long" workers.  Concurrency on the long
worker is always 1, and concurrency on the default one can be much
higher, say 10.