Support multiple apps per environment (#1371)
Goal is to have the same app and utility machines be able to run
different versions of the Class2Go software concurrently.  The different
code versions are called "apps" using the apache terminology.

Taking a modest approach to parameterizing: eventually everything could
be per-app (like database settings, etc.) but for now it's just the
server name (for apache) and the code version.