ExamRecordScoreField.value char-->text (#1449)
This is a little dicey as we've had some bad experiences altering column
types with MySQL, but we're going to give it a go nonetheless.  On
production now we only have 16 rows with data in it, some test data most
likely.

select value,count(*) from c2g_examrecordscorefield where value is not null group by 1;
+-------------------------+----------+
| value                   | count(*) |
+-------------------------+----------+
| [u'b']                  |        4 |
| [u'c']                  |        2 |
| [u'false']              |        3 |
| [u'orthogonal_matrix']  |        1 |
| [u'orthonormal_matrix'] |        1 |
| [u'true']               |        5 |
+-------------------------+----------+