assertEquals was what I wanted, not assertTrue (#1726)
per @jbau review.  I was mistakenly doing assertTrue, which made it not
much of a test at all.  by switching to assertEquals it turns out I
should just be treating the feedback as opaque, which is what I'm doing
now, both cleaner and easier to maintain.  It's never interpreted
anyway...