First-step fix for #1980
Error messages include "Remember: Your username might not be your email
address". Looking up the username requires a separate db lookup and reporting
that the username is wrong leaks security about the presence of a username.
The next step of the fix is changing the user creation/authentication flow a
little.

	modified:   c2g/static/css/core/body.css
	modified:   templates/accounts/changepassword.html
	modified:   templates/accounts/register.html
	modified:   templates/registration/login.html