#2029 - New relation examscore -> examrecordscore
**Note: Migration required for this change.

I'm setting this new relation in 2 instances:

1. Autograded responses. For this the examrecordscore related to is the
one corresponding to the score with penalties excluded.

2. CSV grading. For this, the examrecordscore related to is the one
which is for the latest submission (attempt) at the time of csv upload.
This is because csv grading, as currently works, posts grades to the
last submission for that exam. In this instance, the score on the
examrecirdscore is the same as the examscore as there are no
machine penalties applied with csv grading.

What I have not touched is the regrader. The reason being, that I'm only
setting relations for new submissions, not retroactively. It ought to be
easier to incorporate after current courses have died.