grader cache: switch to db-based cache (#2050)
Expect this to improve hit rates across multiple app servers.  This is a
special operation that we're caching -- infrequent and very expensive.
So it is OK to use a central DB cache for this.  Better would be a
central memcache of course but we don't have one of those (yet).